feat: restrict project update and delete actions to manager

+++ uptask_backend/src/controllers/ProjectController.ts
@@ -73,6 +73,12 @@ export class ProjectController {
                 return res.status(404).json({ error: error.message })
             }
 
+
+            if (project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Solo el Manager puede actualizar el proyecto')
+                return res.status(404).json({error: error.message})
+            }            
+
             project.clientName = req.body.clientName
             project.projectName = req.body.projectName
             project.description = req.body.description
@@ -98,6 +104,11 @@ export class ProjectController {
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({ error: error.message })
             }
+
+            if (project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Solo el Manager puede Eliminar el proyecto')
+                return res.status(404).json({error: error.message})
+            }             
 
             await project.deleteOne()
